refactore: update redux folder name to store name
